backend-drf/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
python-decouple==3.8
reportlab==4.4.4
sqlparse==0.5.3
psycopg2==2.9.11
gunicorn==21.2.0
